+++ ktorfit-ksp/src/main/kotlin/de/jensklingenberg/ktorfit/model/KtorfitClass.kt
@@ -16,5 +16,6 @@ val formParameters = KtorfitClass("", "", "__formParameters")
 val formData = KtorfitClass("", "", "__formData")
 val converterHelper = KtorfitClass("KtorfitConverterHelper", "de.jensklingenberg.ktorfit.internal", "_helper")
 val internalApi = ClassName("de.jensklingenberg.ktorfit.internal", "InternalKtorfitApi")
+val annotationsAttributeKey = KtorfitClass("annotationsAttributeKey", "de.jensklingenberg.ktorfit", "")
 
 fun KtorfitClass.toClassName() = ClassName(packageName, name)