+++ ktorfit-ksp/src/main/kotlin/de/jensklingenberg/ktorfit/poetspec/FunctionSpec.kt
@@ -27,7 +27,8 @@ fun FunctionData.toFunSpec(
             parameterDataList.map {
                 it.parameterSpec()
             },
-        ).addBody(this, resolver, setQualifiedTypeName, returnTypeName)
+        )
+        .addBody(this, resolver, setQualifiedTypeName, returnTypeName)
         .returns(returnTypeName)
         .build()
 }