+++ ktorfit-ksp/src/main/kotlin/de/jensklingenberg/ktorfit/reqBuilderExtension/ReqBuilderExtensionNode.kt
@@ -12,6 +12,7 @@ fun getReqBuilderExtensionText(
     listType: KSType,
     arrayType: KSType,
 ): String {
+    val attributes = getAttributesCode(functionData.parameterDataList, functionData.nonKtorfitAnnotations)
     val method = getMethodCode(functionData.httpMethodAnnotation)
 
     val headers =
@@ -44,14 +45,13 @@ fun getReqBuilderExtensionText(
     val url =
         getUrlCode(functionData.parameterDataList, functionData.httpMethodAnnotation, queryCode)
     val customReqBuilder = getCustomRequestBuilderText(functionData.parameterDataList)
-    val attributeKeys = getAttributeCode(functionData.parameterDataList)
     val args =
         listOf(
+            attributes,
             method,
             url,
             body,
             headers,
-            attributeKeys,
             fields,
             parts,
             customReqBuilder,